# Attached-Assets/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Attendance Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/me/sync:
    post:
      operationId: syncMe
      tags: [users]
      summary: Sync / upsert user record from Clerk identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSync"
      responses:
        "200":
          description: Synced user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Forbidden
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a single user (admin/manager)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Attendance ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, role, name, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [employee, manager, admin]
        department:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserSync:
      type: object
      required: [clerkId, email, name]
      properties:
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [employee, manager, admin]

    # ── Attendance ─────────────────────────────────────────────────────────────
...