# Attached-Assets/lib/api-zod/src/generated/types/userSync.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Attendance Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSync {
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
}